requirements.txt
alembic==1.8.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cycler==0.11.0
Deprecated==1.2.13
dominate==2.6.0
firebase-admin==5.2.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.33.3
gcloud==0.18.3
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.3.1
kiwisolver==1.4.3
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
msgpack==1.0.4
mysql-connector==2.2.9
numpy==1.23.0
oauth2client==4.1.3
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
proto-plus==1.20.6
protobuf==3.20.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pyparsing==3.0.9
Pyrebase==3.0.18
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jwt==3.3.2
pytz==2022.1
requests==2.28.0
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sseclient==0.0.27
uritemplate==4.1.1
urllib3==1.26.9
visitor==0.1.3
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
